Be compatible with tox-venv
- Pass `tryfirst=True` to the hookimpl decorator. This ensures our
plugin runs before tox-venv's does, and since we return non-None,
tox-venv's plugin is not invoked afterwards.
- If tox-venv is installed, call tox-venv's `tox_testenv_create`
directly to create the virtualenv.

Those two changes ensure tox-venv creates the virtualenv only once, and
that it occurs before we install a particular pip version.

Additionally, I reworked the tests and did a little cleanup.

- There's now a `tests/test.py` file that we run with pytest
- `test.py` isolates tests so we can run them in parallel. It creates a
fresh venv in an isolated temporary directory for each test case, and
places the tox workdir in that temporary directory as well.
- Cleanup now unused pieces of the Makefile
- Put dependencies into dev-requirements.txt